Add root makefile to support BSP components
This commit adds a file, 'root.mk' that gets copied as Makefile to the root
of an Android repo checkout as specified in the repo manifest.  It is not
specific to the Overo device and doesn't need to live in the device directory;
this is just a cosy place to store it.

This makes it easy to build the kernel, u-boot, and SGX as part of an Android
build.